+++ class/MailerFactory.class.php
@@ -1,0 +1,8 @@
+<?php
+
+class MailerFactory {
+
+    public function getInstance(){
+        return new Mailer();
+    }
+}