+++ class/actes/ActesNotification.class.php
@@ -2,205 +2,178 @@
 
 class ActesNotification {
 	
-	private $lesFichiers;
 	private $filePath;
-    private $pdfgenerate = false;
-    private $pdftampone;
-    private $newmail=true;
-    private $agent=false;
 
     private $actesTransactionsSQL;
     private $acteTamponne;
     private $authoritySQL;
     private $actesEnveloppeSQL;
 
-    private $sqlQuery;
+    private $mailerFactory;
+
+    private $logger;
 	
 	public function __construct(
 	        ActesTransactionsSQL $actesTransactionsSQL,
             ActeTamponne $acteTamponne,
             AuthoritySQL $authoritySQL,
             ActesEnvelopeSQL $actesEnveloppeSQL,
-            SQLQuery $sqlQuery
+            MailerFactory $mailerFactory,
+            Logger $logger
     ){
-		$this->lesFichiers = array();
 		$this->setFilePath(ACTES_FILES_UPLOAD_ROOT);
 
         $this->actesTransactionsSQL = $actesTransactionsSQL;
         $this->acteTamponne = $acteTamponne;
         $this->authoritySQL = $authoritySQL;
         $this->actesEnveloppeSQL = $actesEnveloppeSQL;
-
-        $this->sqlQuery = $sqlQuery;
+        $this->mailerFactory = $mailerFactory;
+        $this->logger = $logger;
 	}
 	
-	private function setFilePath($filePath){
+	public function setFilePath($filePath){
 		$this->filePath = $filePath;
 	}
 	
 	public function sendAutomaticNotification(){
         foreach($this->actesTransactionsSQL->getTransactionToAutoBroadcast() as $transaction_id){
-            echo "Notification de la transaction $transaction_id\n";
+            $this->log("Notification de la transaction $transaction_id");
             $this->sendNotificationManuel($transaction_id);
         }
 	}
 	
 	public function sendNotificationManuel($transactionId){
-		$transactionInfo = $this->getTransactionInfo($transactionId);
-		if (! $transactionInfo){
-			return false;
-		}
-		$this->sendNotification($transactionInfo);
-		$this->deletedirectoryunzip($transactionInfo);
+        $tmpFolder = new TmpFolder();
+        $tmp_folder = $tmpFolder->create();
+		try {
+            $transactionInfo = $this->actesTransactionsSQL->getInfo($transactionId);
+            if (! $transactionInfo){
+               throw new Exception("La transaction $transactionId n'existe pas");
+            }
+            $this->sendNotification($transactionInfo,$tmp_folder);
+            $tmpFolder->delete($tmp_folder);
+        } catch (Exception $e){
+            $tmpFolder->delete($tmp_folder);
+            throw $e;
+        }
 		return true;
 	}
 	
-	private function sendNotification(array $transaction_info){
+	private function sendNotification(array $transaction_info,$tmp_folder){
         $authority_info = $this->authoritySQL->getInfo($transaction_info['authority_id']);
-
         $envelope_info = $this->actesEnveloppeSQL->getInfo($transaction_info['envelope_id']);
 
 		$defaultBroadcastEmail =explode(',',$authority_info['default_broadcast_email']);
 		$brodcastEmail =explode(',',$transaction_info['broadcast_emails']);
 		$brodcastEmail = array_diff($brodcastEmail,$defaultBroadcastEmail);
 
-        $this->newmail = $authority_info['new_notification'];
-                
+		$fichiers_tamponnees =  $this->tamponnerTGZ($this->filePath . "/" . $envelope_info['file_path'],$transaction_info,$tmp_folder);
+
+
 		if ($transaction_info['auto_broadcasted'] == false){
             //envoie du mail au proprietaire de l'acte
-            $this->agent=true;
-            $this->sendMail($transaction_info,$envelope_info['email'],true);
-            $this->agent=false;
+            $this->sendMail($transaction_info,$envelope_info['email'],true,$authority_info['new_notification'],$fichiers_tamponnees);
             //envoie du mail a toutes les adresses renseignees dans defaut
             foreach($defaultBroadcastEmail as $email){
-                $this->sendMail($transaction_info,$email,true);
+                $this->sendMail($transaction_info,$email,true,false,$fichiers_tamponnees);
             }
             $this->actesTransactionsSQL->setAutoBroadcasted($transaction_info['id']);
 		}
-                
+
 		if ($transaction_info['broadcast_emails']){
             foreach($brodcastEmail as $email){
-                $this->sendMail($transaction_info,$email,$transaction_info['broadcast_send_sources'] == 1);
+                $this->sendMail($transaction_info,$email,$transaction_info['broadcast_send_sources'] == 1,false,$fichiers_tamponnees);
             }
             $this->actesTransactionsSQL->setBroadcasted($transaction_info['id']);
 		}
-                
-        $this->pdfgenerate = false;
 	}
-	
-	private function sendMail($transactionInfo,$emails,$withFile){
+
+	private function sendMail($transactionInfo,$emails,$withFile,$add_url_recup,array $fichiers_tamponnees){
 		if (! $emails){
 			return;
 		}
 
-		$mailer = new Mailer();
-        if(($this->agent && $this->newmail == 'f') || !$this->agent){
-            if(! $this->pdfgenerate){
-                $lesFichiers = $this->getFichiers($transactionInfo);
-            } else {
-                $lesFichiers = $this->pdftampone;
-            }
+		$mailer = $this->mailerFactory->getInstance();
 
-            if ($withFile){
-              foreach ($lesFichiers as $fichier){
-                $mailer->addFile($fichier);
-              }
-            }
-        }
-
-
-        $err = $mailer->addRecipient($emails);
+		$err = $mailer->addRecipient($emails);
         if (! $err){
-                echo "$emails invalide ! \n";
-                return;
+            $this->log("$emails invalide !");
         }
 
-        $trans = new ActesTransaction();
-        $trans->setId($transactionInfo['id']);
-        $trans->init();
-        if($this->newmail == 't') {
-            $mailer->addStringAsFile("retour.xml", $trans->getFluxRetour(4));
+        if($withFile && ! $add_url_recup){
+            foreach($fichiers_tamponnees as $fichier){
+                $mailer->addFile($fichier);
+            }
         }
 
-        $mailContent = $this->getMailContent($transactionInfo);
+        $status_info = $this->actesTransactionsSQL->getStatusInfo($transactionInfo['id'],4);
+
+        $ar_actes_filename = "{$transactionInfo['unique_id']}-{$transactionInfo['type']}-{$transactionInfo['id']}-AR.xml";
+        $mailer->addStringAsFile($ar_actes_filename, $status_info['flux_retour']);
+
+        $mailContent = $this->getMailContent($transactionInfo,$add_url_recup);
 
 		$pdf=new ActesPdf();
 		$pdf->addEmailNotificationField();
 		$pdf->create_pdf($transactionInfo['id']);
 		$monpdf = $pdf->output("bordereau_acquittement","S");
 		$mailer->addStringAsFile("bordereau_acquittement.pdf",$monpdf);
-		
-		$mailer->sendMail("[{$transactionInfo['name']}] Notification d'accusé de réception pour l'acte " . $transactionInfo['number'] , $mailContent);
+
+        $authority_info = $this->authoritySQL->getInfo($transactionInfo['authority_id']);
+
+        $mailer->sendMail(
+                "[{$authority_info['name']}] Notification d'accusé de réception pour l'acte " . $transactionInfo['number'] ,
+                $mailContent
+        );
 	}
-	
-	private function getTransactionInfo($transactionId){
-        return $this->actesTransactionsSQL->getInfo($transactionId);
-	}
-	
-	private function getMailContent($transactionInfo){
-		ob_start();?>
-L'acte de référence interne <?php echo $transactionInfo['number'] ?> a été acquitté sous l'identifiant unique <?php echo $transactionInfo['unique_id']  ?>
+
+	private function getMailContent($transaction_info,$add_url_recup){
+        $status_info = $this->actesTransactionsSQL->getStatusInfo($transaction_info['id'],4);
+        $envelope_info = $this->actesEnveloppeSQL->getInfo($transaction_info['envelope_id']);
 
 
-Nature de l'Acte : <?php echo $transactionInfo['nature_descr'] ?>
+        ob_start();?>
+<?php if ($transaction_info['type'] == 1) : ?>
+L'acte de référence interne <?php echo $transaction_info['number'] ?> a été acquitté sous l'identifiant unique <?php echo $transaction_info['unique_id']  ?>.
+<?php elseif ($transaction_info['type'] == 3) : ?>
+L'envoi de pièces complémentaires (ou du refus explicite) concernant l'actes <?php echo $transaction_info['number'] ?> a été acquitté.
+<?php elseif ($transaction_info['type'] == 4) : ?>
+L'envoi de la lettre d'observation (ou du refus de réponse) concernant l'actes <?php echo $transaction_info['number'] ?> a été acquitté.
+<?php endif; ?>
 
-Objet : <?php echo $transactionInfo['subject'] ?> 
+Nature de l'Acte : <?php echo $transaction_info['nature_descr'] ?>
 
-Décision du : <?php echo $transactionInfo['decision_date']?> 
+Objet : <?php echo $transaction_info['subject'] ?>
 
-Transmise le :  <?php echo $transactionInfo['submission_date']?> 
+Décision du : <?php echo $transaction_info['decision_date']?>
 
-Accusé reçu le :  <?php echo $transactionInfo['date'] ?> 
+Transmise le :  <?php echo $envelope_info['submission_date']?>
 
-<?php
-if($this->agent && $this->newmail == 't'){
-?>
-URL pour récupérer les fichiers : <?php $url = WEBSITE_SSL."/modules/actes/actes_transac_show.php?id=".$transactionInfo['transaction_id']; echo $url; ?>
-<?php }?>
+Accusé reçu le :  <?php echo $status_info['date'] ?>
 
-<?php if($transactionInfo['archive_url']) : ?>
-Archive disponible sur :<?php echo $transactionInfo['archive_url']?>
+<?php if($add_url_recup) : ?>
+URL pour récupérer les fichiers : <?php $url = WEBSITE_SSL."/modules/actes/actes_transac_show.php?id=".$transaction_info['id']; echo $url; ?>
+<?php endif; ?>
+
+<?php if($transaction_info['archive_url']) : ?>
+Archive disponible sur :<?php echo $transaction_info['archive_url']?>
 <?php endif; ?> 
 		<?php 
 		$result = ob_get_contents();
 		ob_end_clean();
 		return $result;
 	}
-	
-	private function getFichiers(array $transactionInfo){
-		$sql = "SELECT file_path " .
-			"FROM actes_transactions LEFT JOIN actes_envelopes " . 
-			"ON actes_transactions.envelope_id=actes_envelopes.id " .
-			"WHERE actes_transactions.id = " .$transactionInfo['id'] ;
-		$result = array();
-		$lesFichiers = $this->sqlQuery->query($sql);
-		foreach($lesFichiers as $f){
-			$result = $this->tamponnerTGZ($this->filePath . "/" . $f['file_path'],$transactionInfo);
-		}
-		
-		$this->pdftampone = $result;
-        $this->pdfgenerate = true;
-		return $result;
-	}
-	
-	private function tamponnerTGZ($filePath,$transactionInfo){
-		
-		$directory_unzip = $filePath."_unzip";
 
-		if (! file_exists($directory_unzip)){
-			mkdir($directory_unzip);
-		}
-		//FIXME beuark !
-		$cmd = "tar xzf $filePath  -C ".$directory_unzip;
-		Trace::wrap_exec($cmd, $status, $ret);
-		chdir($directory_unzip);
-		
+	
+	private function tamponnerTGZ($filePath,$transactionInfo,$tmp_folder){
+        $pharData = new PharData($filePath);
+        $pharData->extractTo($tmp_folder);
+
 		$result = array();
-		
-		$files = scandir($directory_unzip);
+		$files = array_diff(scandir($tmp_folder),array('.','..'));
 		foreach($files as $file){
-			$path_parts = pathinfo($file);
-			if ($path_parts['extension'] == 'pdf'){
+		    $file = $tmp_folder."/$file";
+			if (pathinfo($file,PATHINFO_EXTENSION) == 'pdf'){
 				$fileString = $this->acteTamponne->tamponnerPDF($file,$transactionInfo['id']);
 				file_put_contents($file,$fileString);
 			}
@@ -209,28 +182,7 @@ Archive disponible sur :<?php echo $transactionInfo['archive_url']?>
 		return $result;
 	}
 
-	
-	/**
-	 * La fonction tamponnerTGZ extrait l'enveloppe tgz envoyée au MIOCT
-	 * pour récupérer les pdf et les tamponner. Cette extraction n'est plus utile par 
-	 * la suite. Il faut donc supprimer le dossier.
-	 * @param $transactionInfo array
-	 */
-	private function deletedirectoryunzip($transactionInfo){
-		chdir(TEDETIS_TMP_PATH);
-		$sql = "SELECT file_path " .
-			"FROM actes_transactions LEFT JOIN actes_envelopes " . 
-			"ON actes_transactions.envelope_id=actes_envelopes.id " .
-			"WHERE actes_transactions.id = " .$transactionInfo['id'] ;
-
-		$listefichiers = $this->sqlQuery->query($sql);
-		foreach($listefichiers as $f){
-			$directory_unzip = $this->filePath . "/" . $f['file_path']."_unzip";
-			if (file_exists($directory_unzip)){
-				$cmd = "rm -r $directory_unzip";
-                Trace::wrap_exec($cmd, $status, $ret);	
-			}
-		}
-	}
-
+    private function log($message){
+        $this->logger->log("actes-notification",$message);
+    }
 }