+++ class/actes/ActesPdf.class.php
@@ -1,5 +1,6 @@
 <?php 
 
+require_once __DIR__."/../../public.ssl/modules/actes/class/ActesTransaction.class.php";
 
 class ActesPdf {