+++ public.ssl/modules/actes/actes_transac_reponse_create.php
@@ -115,6 +115,7 @@ $trans->set("user_id",$related_trans->get('user_id'));
 $trans->set("authority_id",$related_trans->get('authority_id'));
 $trans->set("classification",$related_trans->get('classification'));
 $trans->set("classification_date",$related_trans->get('classification_date'));
+$trans->set("unique_id",$related_trans->get('unique_id'));
 
 
 // Destination de création des fichiers